+++ packages/server/dataloader/primaryKeyLoaderMakers.ts
@@ -6,6 +6,7 @@ import getMeetingTemplatesByIds from '../postgres/queries/getMeetingTemplatesByI
 import getTemplateRefsByIds from '../postgres/queries/getTemplateRefsByIds'
 import {getUsersByIds} from '../postgres/queries/getUsersByIds'
 import {
+  selectAgendaItems,
   selectMeetingSettings,
   selectOrganizations,
   selectRetroReflections,
@@ -87,6 +88,10 @@ export const suggestedActions = primaryKeyLoaderMaker((ids: readonly string[]) =
   return selectSuggestedAction().where('id', 'in', ids).execute()
 })
 
-export const _PGmeetingSettings = primaryKeyLoaderMaker((ids: readonly string[]) => {
+export const meetingSettings = primaryKeyLoaderMaker((ids: readonly string[]) => {
   return selectMeetingSettings().where('id', 'in', ids).execute()
+})
+
+export const _pgagendaItems = primaryKeyLoaderMaker((ids: readonly string[]) => {
+  return selectAgendaItems().where('id', 'in', ids).execute()
 })