+++ packages/server/graphql/types/PokerMeetingSettings.ts
@@ -1,8 +1,0 @@
-import {GraphQLObjectType} from 'graphql'
-
-const PokerMeetingSettings = new GraphQLObjectType({
-  name: 'PokerMeetingSettings',
-  fields: {}
-})
-
-export default PokerMeetingSettings